ref(js): Extract codeKeywords from codeBlock component
